:sparkles: refacto of some files
